wip inheritance method for modularizing authentication
